% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_annotate.R
\name{gd_anotate}
\alias{gd_anotate}
\title{gff_annotate() returns annotation from a list of mutation from df}
\usage{
gd_anotate(gff_file, subtracted)
}
\arguments{
\item{gff_file}{path to gff3 reference file from breseq/data}

\item{subtracted}{comes from gd_substract()}
}
\value{
a data frame containing for each mutation a row with anotation
}
\description{
gff_annotate() returns annotation from a list of mutation from df
}
