% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_load.R
\name{gd_load}
\alias{gd_load}
\title{gd_load() import a gd (genome diff) file in a data frame}
\usage{
gd_load(gd_file)
}
\arguments{
\item{gd_file}{path to the gd file}
}
\value{
a dataframe
}
\description{
gd_load() import a gd (genome diff) file in a data frame
}
