% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_merge.R
\name{gd_merge}
\alias{gd_merge}
\title{gd_merge() call gd_read() on gd files and merge them in an unique data frame with an origin column}
\usage{
gd_merge(gd_list)
}
\arguments{
\item{gd_list}{a list of path to gd file}
}
\value{
a data frame containing each mutation from each gd files indexed by a new origin col
}
\description{
gd_merge() call gd_read() on gd files and merge them in an unique data frame with an origin column
}
