% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_subtract.R
\name{gd_subtract}
\alias{gd_subtract}
\title{gd_subtract() remove mutation present from reference in other gd file an output it as data frame}
\usage{
gd_subtract(gd_merged, ref_name, filtered_out = TRUE)
}
\arguments{
\item{gd_merged}{a dataframe from gd_merged()}

\item{ref_name}{name (with .gd) of the reference gd file}

\item{filtered_out}{remove row in ref. Default = TRUE}
}
\value{
a dataframe
}
\description{
gd_subtract() remove mutation present from reference in other gd file an output it as data frame
}
