% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_generator.R
\name{tree_generator}
\alias{tree_generator}
\title{Perform neighbour joining (perform a distance matrix and sample matrix, then use APE)}
\usage{
tree_generator(annotated_mutations)
}
\arguments{
\item{annotated_mutations}{result from gd_annotate()}
}
\value{
a phylo object
}
\description{
Perform neighbour joining (perform a distance matrix and sample matrix, then use APE)
}
